1/ Add option --enable_macsec 2/ Refact macsec fixtures
